// frontend/src_backup/shared/types/api.ts
export interface UserProfile {
  id: number;
  username: string;
  email: string;
  name?: string;
  role: string;
  is_active: boolean;
  is_superuser: boolean;
  email_verified: boolean;
  email_verified_at?: string;
  last_login?: string;
  permissions?: string[];
  
  // UI расширения (не в бэкенде)
  avatar?: string;
  first_name?: string;
  last_name?: string;
}